Move zoom and projection functions into DiagramHelper model
