+++ src/diagram/index.ts
@@ -1,5 +1,6 @@
 import { Diagram } from "./components/diagram";
 import { EditVariableDialogContent, updateVariable } from "./components/dialog/edit-variable-dialog";
+import { DiagramHelperType } from "./models/diagram-helper";
 import { DQNode } from "./models/dq-node";
 import { DQRoot } from "./models/dq-root";
 import { Variable, VariableType } from "./models/variable";
@@ -9,4 +10,7 @@ import { useSelectMultipleVariables } from "../hooks/use-select-multiple-variabl
 
 import "./components/diagram.scss";
 
-export { Diagram, EditVariableDialogContent, updateVariable, DQNode, DQRoot, useSelectMultipleVariables, Variable, VariableChip, VariableChipList, VariableType};
+export {
+  Diagram, DiagramHelperType, EditVariableDialogContent, updateVariable, DQNode, DQRoot, useSelectMultipleVariables,
+  Variable, VariableChip, VariableChipList, VariableType
+};
